# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Second-Hand Store API",
    "description": "Fully autonomous second-hand store with AI negotiation",
    "version": "1.0.0"
  },
  "paths": {
...
    "/admin/orders/{order_id}": {
      "get": {
        "tags": [
          "System"
        ],
        "summary": "Get Order",
        "description": "Get a specific order by ID.",
        "operationId": "get_order_admin_orders__order_id__get",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "System"
        ],
        "summary": "Update Order",
        "description": "Update order details.",
        "operationId": "update_order_admin_orders__order_id__put",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Order Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "System"
        ],
        "summary": "Delete Order",
        "description": "Delete an order.",
        "operationId": "delete_order_admin_orders__order_id__delete",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/orders/{order_id}/status": {
      "put": {
        "tags": [
          "System"
        ],
        "summary": "Update Order Status",
        "description": "Update order status. Reaching 'delivered' also tells the buyer (SPEC-057).",
        "operationId": "update_order_status_admin_orders__order_id__status_put",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Order Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderStatusUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/orders/{order_id}/shipment": {
      "put": {
        "tags": [
          "System"
        ],
        "summary": "Record Shipment",
        "description": "Record postage for a paid order and tell the buyer about it.\n\nSPEC-057. Before this the lifecycle simply stopped: `status` could be set to\n'shipped' but there was nowhere to put what it shipped WITH, so the seller\npasted tracking numbers into the chat by hand and the agent could only\nrepeat the word \"shipped\" when asked.\n\nOne call does all four things \u2014 record, stamp, email, post to the chat \u2014\nbecause a seller who has to remember the other three will eventually not.",
        "operationId": "record_shipment_admin_orders__order_id__shipment_put",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Order Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShipmentUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/orders/refund/{item_id}": {
      "post": {
        "tags": [
          "System"
        ],
        "summary": "Refund Item",
        "description": "Refund a paid item, releasing it back to the catalogue.\n\nSPEC-056 #1. This used to be `POST /payment/refund/{item_id}` on the public\npayment router with `Depends(verify_admin)` and nothing more. Admin auth is\na cookie the browser attaches by itself, which is exactly the shape CSRF\nexploits \u2014 every other mutating admin action is gated by `verify_csrf_token`\nvia the `protected` router in `routes/admin/__init__.py`, and the one that\nmoves money was the exception. Living here, it inherits that gate.",
        "operationId": "refund_item_admin_orders_refund__item_id__post",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          },
          {
            "name": "reason",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Reason"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OrderStatusUpdate": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "OrderStatusUpdate"
      },
      "OrderUpdate": {
        "properties": {
          "item_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Item Name"
          },
          "amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "shipping_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shipping Address"
          },
          "shipping_phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shipping Phone"
          },
          "shipping_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shipping Name"
          },
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "recipient_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recipient Name"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "title": "OrderUpdate"
      },
...
      "ShipmentUpdate": {
        "properties": {
          "courier": {
            "type": "string",
            "title": "Courier"
          },
          "tracking_number": {
            "type": "string",
            "title": "Tracking Number"
          },
          "tracking_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tracking Url"
          },
          "notify": {
            "type": "boolean",
            "title": "Notify",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "courier",
          "tracking_number"
        ],
        "title": "ShipmentUpdate",
        "description": "Postage as the seller records it (SPEC-057).\n\n`tracking_url` is optional because it is usually derivable from the courier;\n`notify` exists so a correction to an already-announced shipment doesn't\nemail the buyer a second time."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}